Move GridIndex to its own file
